Zmlp/zmlp 377 emit task progress/status

See merge request zorroa-zvi/zmlp!811

+++ services/archivist/archivist/src/main/kotlin/com/zorroa/archivist/domain/Task.kt
@@ -133,7 +133,7 @@ open class Task(
     val assetCounts: Map<String, Int>,
 
     @ApiModelProperty("Task Progress")
-    var progress: Short = 0,
+    var progress: Int = 0,
 
     @ApiModelProperty("Current Task Status")
     var status: String? = null
